Gcc13

+++ Libraries/Aurora/Source/MaterialDefinition.cpp
@@ -10,6 +10,7 @@
 #include "pch.h"
 
 #include "MaterialDefinition.h"
+#include <vector>
 
 BEGIN_AURORA
 
@@ -19,7 +20,7 @@ MaterialDefaultValues::MaterialDefaultValues(const UniformBufferDefinition& prop
 {
     AU_ASSERT(
         defaultProps.size() == propertyDefs.size(), "Default properties do not match definition");
-    for (int i = 0; i < defaultTxt.size(); i++)
+    for (size_t i = 0; i < defaultTxt.size(); i++)
     {
         textureNames.push_back(defaultTxt[i].name);
     }
